requirements.txt
mediapipe==0.10.5
opencv-python==4.11.0.86
pyserial==3.5
